rename page folder
