+++ src/app/authentication/authentication-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
  
-import { LoginComponent } from './screens';
+import { LoginComponent } from './pages';
  
 const routes: Routes = [
   { path: '', component: LoginComponent }
