+++ src/app/authentication/authentication.module.ts
@@ -1,7 +1,6 @@
-
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { AuthenticationRoutingModule } from './authentication-routing.module';
-import { LoginComponent } from './screens';
+import { LoginComponent } from './pages';
 import { FormLoginComponent } from './components/form-login/form-login.component';
 
 @NgModule({
